Fixing some warnings found by Xcode

+++ meteoio/plugins/CsvIO.h
@@ -62,7 +62,7 @@ class CsvParameters {
 		std::string file_and_path, datetime_format, time_format, single_field; 		///< the scanf() format string for use in parseDate, the parameter in case of a single value contained in the Csv file
 		std::string name, id;
 		double slope, azi;
-		float csv_tz;		///< timezone to apply to parsed dates
+		double csv_tz;		///< timezone to apply to parsed dates
 		bool has_tz;		///< does the user-provided date/time format contains a TZ?
 };
